chore: get getNonFungibleTokenInfo working

+++ examples/src/main/resources/precompile-example/HederaTokenService.sol
@@ -156,8 +156,8 @@ abstract contract HederaTokenService is HederaResponseCodes {
     /// @return tokenAddress the created token's address
     function createFungibleToken(
         IHederaTokenService.HederaToken memory token,
-        uint initialTotalSupply,
-        uint decimals) nonEmptyExpiry(token)
+        uint64 initialTotalSupply,
+        uint32 decimals) nonEmptyExpiry(token)
     internal returns (int responseCode, address tokenAddress) {
 
         (bool success, bytes memory result) = precompileAddress.call{value: msg.value}(
@@ -179,8 +179,8 @@ abstract contract HederaTokenService is HederaResponseCodes {
     /// @return tokenAddress the created token's address
     function createFungibleTokenWithCustomFees(
         IHederaTokenService.HederaToken memory token,
-        uint initialTotalSupply,
-        uint decimals,
+        uint64 initialTotalSupply,
+        uint32 decimals,
         IHederaTokenService.FixedFee[] memory fixedFees,
         IHederaTokenService.FractionalFee[] memory fractionalFees) nonEmptyExpiry(token)
     internal returns (int responseCode, address tokenAddress) {
@@ -617,6 +617,7 @@ abstract contract HederaTokenService is HederaResponseCodes {
 
     /// Operation to update token expiry info
     /// @param token The token address
+    /// @param expiryInfo The expiry properties of a token
     /// @return responseCode The response code for the status of the request. SUCCESS is 22.
     function updateTokenExpiryInfo(address token, IHederaTokenService.Expiry memory expiryInfo) external returns (int responseCode){
         (bool success, bytes memory result) = precompileAddress.call(
@@ -624,9 +625,10 @@ abstract contract HederaTokenService is HederaResponseCodes {
         (responseCode) = success ? abi.decode(result, (int32)) : HederaResponseCodes.UNKNOWN;
     }
 
-    /// Operation to update token expiry info
+    /// Operation to update specified token's keys
+    /// The key type to update is specified in the keys parameter
     /// @param token The token address
-    /// @param keys The token keys
+    /// @param keys The token key type to update
     /// @return responseCode The response code for the status of the request. SUCCESS is 22.
     function updateTokenKeys(address token, IHederaTokenService.TokenKey[] memory keys) external returns (int64 responseCode){
         (bool success, bytes memory result) = precompileAddress.call(
